requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-dotenv==1.0.0
starlette==0.27.0
sqlalchemy==2.0.15
toml==0.10.2
# psycopg2==2.9.7
python-jose==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.11.1
# pydantic==2.7.0
psycopg2==2.9.7
psycopg2-binary== 2.9.9
openai==0.27.10
pydantic[email]
